requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
prisma>=0.11.0
supabase>=1.0.0
python-dotenv>=0.21.0
Jinja2>=3.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0
aiofiles>=23.0.0
python-multipart>=0.0.6